fix(mcp): use canonical JSONL stdio framing
