+++ Sources/QuillCodeTools/MCPStdioMessageCodec.swift
@@ -3,25 +3,74 @@ import Foundation
 public enum MCPStdioMessageCodec {
     public static let maxMessageBytes = 5_000_000
 
+    private static let legacyHeaderPrefix = Data("Content-Length:".utf8)
+    private static let legacyHeaderSeparator = Data("\r\n\r\n".utf8)
+
     public static func encodeJSONObject(_ object: [String: Any]) throws -> Data {
-        let header = "Content-Length: \(try jsonBody(object).count)\r\n\r\n"
-        var data = Data(header.utf8)
-        data.append(try jsonBody(object))
+        var data = try jsonBody(object)
+        guard data.count <= maxMessageBytes else {
+            throw MCPProbeError.invalidMessage("MCP message exceeded \(maxMessageBytes) bytes.")
+        }
+        data.append(0x0A)
         return data
     }
 
-    /// Serialize a JSON-RPC message to bare JSON bytes, WITHOUT the stdio `Content-Length`
-    /// framing header. Used by the HTTP transport, whose framing is HTTP itself.
+    /// Serialize a JSON-RPC message to bare JSON bytes. The HTTP transport supplies its own framing;
+    /// the stdio transport appends one newline through `encodeJSONObject` as required by MCP.
     public static func jsonBody(_ object: [String: Any]) throws -> Data {
         try JSONSerialization.data(withJSONObject: object, options: [])
     }
 
     public static func nextMessageData(from buffer: inout Data) throws -> Data? {
-        let separator = Data("\r\n\r\n".utf8)
-        guard let headerRange = buffer.range(of: separator) else {
-            return nil
-        }
+        while !buffer.isEmpty {
+            if buffer.starts(with: legacyHeaderPrefix) {
+                return try nextLegacyMessageData(from: &buffer)
+            }
+            if legacyHeaderPrefix.starts(with: buffer) {
+                return nil
+            }
 
+            guard let newline = buffer.firstIndex(of: 0x0A) else {
+                guard buffer.count <= maxMessageBytes else {
+                    throw MCPProbeError.invalidMessage("MCP message exceeded \(maxMessageBytes) bytes.")
+                }
+                return nil
+            }
+
+            var message = buffer.subdata(in: buffer.startIndex..<newline)
+            buffer.removeSubrange(buffer.startIndex...newline)
+            if message.last == 0x0D { message.removeLast() }
+            if message.isEmpty { continue }
+            guard message.count <= maxMessageBytes else {
+                throw MCPProbeError.invalidMessage("MCP message exceeded \(maxMessageBytes) bytes.")
+            }
+            return message
+        }
+        return nil
+    }
+
+    /// Accept the framing used by early QuillCode builds so an upgraded client can still connect to
+    /// a legacy local server. New outbound messages always use MCP's newline-delimited JSON format.
+    static func encodeLegacyJSONObject(_ object: [String: Any]) throws -> Data {
+        let body = try jsonBody(object)
+        guard body.count <= maxMessageBytes else {
+            throw MCPProbeError.invalidMessage("MCP message exceeded \(maxMessageBytes) bytes.")
+        }
+        var data = Data("Content-Length: \(body.count)\r\n\r\n".utf8)
+        data.append(body)
+        return data
+    }
+
+    public static func decodeJSONObject(_ data: Data) throws -> [String: Any] {
+        let object = try JSONSerialization.jsonObject(with: data, options: [])
+        guard let dictionary = object as? [String: Any] else {
+            throw MCPProbeError.invalidMessage("MCP message body is not a JSON object.")
+        }
+        return dictionary
+    }
+
+    private static func nextLegacyMessageData(from buffer: inout Data) throws -> Data? {
+        guard let headerRange = buffer.range(of: legacyHeaderSeparator) else { return nil }
         let headerData = buffer.subdata(in: buffer.startIndex..<headerRange.lowerBound)
         guard let header = String(data: headerData, encoding: .utf8) else {
             throw MCPProbeError.invalidMessage("MCP message header is not UTF-8.")
@@ -40,14 +89,6 @@ public enum MCPStdioMessageCodec {
         let message = buffer.subdata(in: bodyStart..<bodyEnd)
         buffer.removeSubrange(buffer.startIndex..<bodyEnd)
         return message
-    }
-
-    public static func decodeJSONObject(_ data: Data) throws -> [String: Any] {
-        let object = try JSONSerialization.jsonObject(with: data, options: [])
-        guard let dictionary = object as? [String: Any] else {
-            throw MCPProbeError.invalidMessage("MCP message body is not a JSON object.")
-        }
-        return dictionary
     }
 
     private static func contentLength(from header: String) throws -> Int {
